profile, bodymass, dan reference

+++ src/project/trackfit/model/User.java
@@ -6,9 +6,9 @@ public class User {
 	private int height;
 	private int weight;
 	private int age;
-	private int body_mass;
+	private double body_mass;
 	
-	public User(int id, String name, int height, int weight, int age, int body_mass) {
+	public User(int id, String name, int height, int weight, int age, double body_mass) {
 		this.id = id;
 		this.name = name;
 		this.height = height;
@@ -37,7 +37,7 @@ public class User {
 		return age;
 	}
 	
-	public int getBodyMass() {
+	public double getBodyMass() {
 		return body_mass;
 	}
 	
@@ -57,7 +57,7 @@ public class User {
 		this.age = age;
 	}
 	
-	public void setBodyMass(int body_mass) {
+	public void setBodyMass(double body_mass) {
 		this.body_mass = body_mass;
 	}
 }
